1742 Argument check

+++ cleanup-junk-files.sh
@@ -1,5 +1,7 @@
 #!/bin/zsh
 
+test -n "$1" || exit 1
+
 rm -vf $1/**/._*
 rm -vf $1/**/.DS_Store
 rm -vf $1/**/.*.un\~
